Fix Merge conflict issue
