+++ app/src/main/java/com/example/mesh_base/global_interfaces/ConnectionHandlersEnum.java
@@ -1,0 +1,6 @@
+package com.example.mesh_base.global_interfaces;
+
+public enum ConnectionHandlersEnum {
+    BLE,
+    WifiDirect
+}
