+++ app/src/main/java/com/example/mesh_base/mesh_manager/MeshManagerListener.java
@@ -4,63 +4,48 @@ import android.util.Log;
 
 import com.example.mesh_base.global_interfaces.Device;
 
-import java.util.ArrayList;
-
 public abstract class MeshManagerListener {
-  String TAG = "my_meshManager_listener";
+    String TAG = "my_meshManager_listener";
 
-  static MeshManagerListener createEmpty() {
-    return new MeshManagerListener() {
-      @Override
-      public void onData(byte[] data, Device device) {
-        Log.d(TAG, "[from empty listener] received data");
-      }
+    static MeshManagerListener createEmpty() {
+        return new MeshManagerListener() {
+            @Override
+            public void onDataReceivedForSelf(byte[] data) {
+                Log.d(TAG, "[from empty listener] received data");
+            }
 
-      @Override
-      public void onStatusChange(Status status) {
-        Log.d(TAG, "[from empty listener] status changed");
-      }
+            @Override
+            public void onStatusChange(Status status) {
+                Log.d(TAG, "[from empty listener] status changed");
+            }
 
-      @Override
-      public void onNeighborsChanged(ArrayList<Device> devices) {
-        Log.d(TAG, "[from empty listener] neighbors changed");
-      }
+            @Override
+            public void onNeighborConnected(Device device) {
+                Log.d(TAG, "[from empty listener] device connected");
+            }
 
-      @Override
-      public void onConnected(Device device) {
-        Log.d(TAG, "[from empty listener] device connected");
-      }
+            @Override
+            public void onNeighborDisconnected(Device device) {
+                Log.d(TAG, "[from empty listener] device disconnected");
+            }
 
-      @Override
-      public void onDisconnected(Device device) {
-        Log.d(TAG, "[from empty listener] device disconnected");
-      }
+            @Override
+            public void onError(Exception e) {
+                Log.d(TAG, "[from empty listener] error");
+            }
+        };
 
-      @Override
-      public void onDiscovered(Device device) {
-        Log.d(TAG, "[from empty listener] device discovered");
-      }
+    }
 
-      @Override
-      public void onError(Exception e) {
-        Log.d(TAG, "[from empty listener] error");
-      }
-    };
+    // Is going to be called only when the received data destination matches the current
+    // user UUID, In other words if we are the destination.
+    abstract public void onDataReceivedForSelf(byte[] data);
 
-  }
+    abstract public void onStatusChange(Status status);
 
-  abstract public void onData(byte[] data, Device device);
+    abstract public void onNeighborConnected(Device device);
 
-  abstract public void onStatusChange(Status status);
+    abstract public void onNeighborDisconnected(Device device);
 
-  //TODO: add medium type in devices
-  abstract public void onNeighborsChanged(ArrayList<Device> devices);
-
-  abstract public void onConnected(Device device);
-
-  abstract public void onDisconnected(Device device);
-
-  abstract public void onDiscovered(Device device);
-
-  abstract public void onError(Exception e);
+    abstract public void onError(Exception e);
 }
