+++ app/src/main/java/com/example/mesh_base/mesh_manager/Status.java
@@ -1,52 +1,61 @@
 package com.example.mesh_base.mesh_manager;
 
+import com.example.mesh_base.global_interfaces.ConnectionHandlersEnum;
+
+import java.util.HashMap;
+
 public class Status {
-  private final Property ble;
-  private final Property wifiDirect;
-
-  private final boolean isOn;
-
-  Status(boolean isOn, Property ble, Property wifiDirect) {
-    this.isOn = isOn;
-    this.ble = ble;
-    this.wifiDirect = wifiDirect;
-  }
-
-  public Property getBle() {
-    return ble;
-  }
-
-  public Property getWifiDirect() {
-    return wifiDirect;
-  }
-
-  public boolean isOn() {
-    return isOn;
-  }
-
-  public static class Property {
-
-    private final boolean isSupported;
     private final boolean isOn;
-    private final boolean isAllowed;
+    private final HashMap<ConnectionHandlersEnum, Property> connectionStatuses;
 
-    public Property(boolean isSupported, boolean isOn, boolean isAllowed) {
-      this.isSupported = isSupported;
-      this.isOn = isOn;
-      this.isAllowed = isAllowed;
+    Status(boolean isOn, HashMap<ConnectionHandlersEnum, Property> connections) {
+        this.isOn = isOn;
+        this.connectionStatuses = connections;
     }
 
-    public boolean isSupported() {
-      return isSupported;
+    public HashMap<ConnectionHandlersEnum, Property> getStatus() {
+        return connectionStatuses;
     }
+
+    public Property getStatus(ConnectionHandlersEnum connectionType) {
+        return connectionStatuses.get(connectionType);
+    }
+
+//    public Property getBle() {
+//        return ble;
+//    }
+//
+//    public Property getWifiDirect() {
+//        return wifiDirect;
+//    }
 
     public boolean isOn() {
-      return isOn;
+        return isOn;
     }
 
-    public boolean isAllowed() {
-      return isAllowed;
+    public static class Property {
+
+        private final boolean isSupported;
+        private final boolean isOn;
+        private final boolean isAllowed;
+
+        public Property(boolean isSupported, boolean isOn, boolean isAllowed) {
+            this.isSupported = isSupported;
+            this.isOn = isOn;
+            this.isAllowed = isAllowed;
+        }
+
+        public boolean isSupported() {
+            return isSupported;
+        }
+
+        public boolean isOn() {
+            return isOn;
+        }
+
+        public boolean isAllowed() {
+            return isAllowed;
+        }
     }
-  }
 
 }
