+++ gradle/libs.versions.toml
@@ -1,5 +1,5 @@
 [versions]
-agp = "8.7.2"
+agp = "8.5.0"
 core = "1.6.1"
 kotlin = "2.0.0"
 coreKtx = "1.15.0"
